# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Email Campaign Platform API
  version: 1.0.0
  description: API surface for admin workflows, campaign sending, tracking, unsubscribe, and SES event handling.
servers:
  - url: https://email-campaign-platform.onrender.com
    description: Production API
  - url: http://localhost:4000
    description: Local API
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/lists/{id}:
    get:
      summary: Get one list with stats
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List detail
    put:
      summary: Update a contact list
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated
    delete:
      summary: Delete a contact list
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
  /api/lists/{id}/contacts:
    get:
      summary: Get contacts in a list
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Paginated contacts
  /api/lists/{id}/contacts.csv:
    get:
      summary: Export contacts in a list as CSV
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: CSV export
  /api/lists/{id}/import:
    post:
      summary: Preview or commit a CSV contact import
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Preview rows or import result
...